Extract NDBaseInterfaceOrchestrator from LoopbackInterfaceOrchestrator
Create an intermediate base class for interface orchestrators that holds
shared lifecycle methods (deploy_pending, remove_pending, queue management,
FabricContext integration). This avoids modifying NDBaseOrchestrator and
resolves Pylance type errors in nd_interface_loopback.py by narrowing the
model_orchestrator type via assert isinstance().
