+++ plugins/module_utils/orchestrators/loopback_interface.py
@@ -25,24 +25,22 @@ from typing import ClassVar, List, Optional, Type
 
 from ansible_collections.cisco.nd.plugins.module_utils.endpoints.base import NDEndpointBaseModel
 from ansible_collections.cisco.nd.plugins.module_utils.endpoints.v1.manage.manage_interfaces import (
-    EpManageInterfacesDeploy,
     EpManageInterfacesGet,
     EpManageInterfacesListGet,
     EpManageInterfacesPost,
     EpManageInterfacesPut,
     EpManageInterfacesRemove,
 )
-from ansible_collections.cisco.nd.plugins.module_utils.fabric_context import FabricContext
 from ansible_collections.cisco.nd.plugins.module_utils.models.base import NDBaseModel
 from ansible_collections.cisco.nd.plugins.module_utils.models.interfaces.loopback_interface import (
     LOOPBACK_POLICY_TYPE_MAPPING,
     LoopbackInterfaceModel,
 )
-from ansible_collections.cisco.nd.plugins.module_utils.orchestrators.base import NDBaseOrchestrator
+from ansible_collections.cisco.nd.plugins.module_utils.orchestrators.base_interface import NDBaseInterfaceOrchestrator
 from ansible_collections.cisco.nd.plugins.module_utils.orchestrators.types import ResponseType
 
 
-class LoopbackInterfaceOrchestrator(NDBaseOrchestrator[LoopbackInterfaceModel]):
+class LoopbackInterfaceOrchestrator(NDBaseInterfaceOrchestrator[LoopbackInterfaceModel]):
     """
     # Summary
 
@@ -87,194 +85,6 @@ class LoopbackInterfaceOrchestrator(NDBaseOrchestrator[LoopbackInterfaceModel]):
     query_all_endpoint: Type[NDEndpointBaseModel] = EpManageInterfacesListGet
     create_bulk_endpoint: Type[NDEndpointBaseModel] = EpManageInterfacesPost
     delete_bulk_endpoint: Type[NDEndpointBaseModel] = EpManageInterfacesRemove
-
-    deploy: bool = True
-
-    _fabric_context: Optional[FabricContext] = None
-    _pending_deploys: list[tuple[str, str]] = []
-    _pending_removes: list[tuple[str, str]] = []
-
-    @property
-    def fabric_name(self) -> str:
-        """
-        # Summary
-
-        Return `fabric_name` from module params.
-
-        ## Raises
-
-        None
-        """
-        return self.sender.params.get("fabric_name")
-
-    @property
-    def fabric_context(self) -> FabricContext:
-        """
-        # Summary
-
-        Return a lazily-initialized `FabricContext` for this orchestrator's fabric.
-
-        ## Raises
-
-        None
-        """
-        if self._fabric_context is None:
-            self._fabric_context = FabricContext(sender=self.sender, fabric_name=self.fabric_name)
-        return self._fabric_context
-
-    def _resolve_switch_id(self, switch_ip: str) -> str:
-        """
-        # Summary
-
-        Resolve a `switch_ip` to its `switchId` via `FabricContext`.
-
-        ## Raises
-
-        ### RuntimeError
-
-        - If no switch matches the given IP in the fabric.
-        """
-        return self.fabric_context.get_switch_id(switch_ip)
-
-    def validate_prerequisites(self) -> None:
-        """
-        # Summary
-
-        Run pre-flight validation before any CRUD operations. Checks that the fabric exists and is modifiable.
-
-        ## Raises
-
-        ### RuntimeError
-
-        - If the fabric does not exist on the target ND node.
-        - If the fabric is in deployment-freeze mode.
-        """
-        self.fabric_context.validate_for_mutation()
-
-    def _configure_endpoint(self, api_endpoint, switch_sn: str):
-        """
-        # Summary
-
-        Set `fabric_name` and `switch_sn` on an endpoint instance before path generation.
-
-        ## Raises
-
-        None
-        """
-        api_endpoint.fabric_name = self.fabric_name
-        api_endpoint.switch_sn = switch_sn
-        return api_endpoint
-
-    def _queue_deploy(self, interface_name: str, switch_id: str) -> None:
-        """
-        # Summary
-
-        Queue an `(interface_name, switch_id)` pair for deferred deployment. Call `deploy_pending` after all mutations
-        are complete to deploy in bulk.
-
-        ## Raises
-
-        None
-        """
-        pair = (interface_name, switch_id)
-        if pair not in self._pending_deploys:
-            self._pending_deploys.append(pair)
-
-    def _queue_remove(self, interface_name: str, switch_id: str) -> None:
-        """
-        # Summary
-
-        Queue an `(interface_name, switch_id)` pair for deferred bulk removal. Call `remove_pending` after all mutations
-        are complete to remove in bulk.
-
-        ## Raises
-
-        None
-        """
-        pair = (interface_name, switch_id)
-        if pair not in self._pending_removes:
-            self._pending_removes.append(pair)
-
-    def deploy_pending(self) -> ResponseType | None:
-        """
-        # Summary
-
-        Deploy all queued interface configurations in a single API call via `interfaceActions/deploy`. Clears the pending
-        queue after deployment.
-
-        When `deploy` is `False`, returns `None` without making any API call.
-
-        ## Raises
-
-        ### RuntimeError
-
-        - If the deploy API request fails.
-        """
-        if not self.deploy or not self._pending_deploys:
-            return None
-        try:
-            result = self._deploy_interfaces()
-            self._pending_deploys = []
-            return result
-        except Exception as e:
-            raise RuntimeError(f"Bulk deploy failed for interfaces {self._pending_deploys}: {e}") from e
-
-    def _deploy_interfaces(self) -> ResponseType:
-        """
-        # Summary
-
-        Deploy queued interfaces via `interfaceActions/deploy`. Sends the explicit list of `{interfaceName, switchId}` pairs.
-
-        ## Raises
-
-        ### Exception
-
-        - If the deploy API request fails (propagated to caller).
-        """
-        api_endpoint = EpManageInterfacesDeploy()
-        api_endpoint.fabric_name = self.fabric_name
-        payload = {"interfaces": [{"interfaceName": name, "switchId": switch_id} for name, switch_id in self._pending_deploys]}
-        return self.sender.request(path=api_endpoint.path, method=api_endpoint.verb, data=payload)
-
-    def remove_pending(self) -> ResponseType | None:
-        """
-        # Summary
-
-        Remove all queued interfaces in a single API call via `interfaceActions/remove`. Clears the pending queue after removal.
-
-        Returns `None` without making any API call if the queue is empty.
-
-        ## Raises
-
-        ### RuntimeError
-
-        - If the remove API request fails.
-        """
-        if not self._pending_removes:
-            return None
-        try:
-            result = self._remove_interfaces()
-            self._pending_removes = []
-            return result
-        except Exception as e:
-            raise RuntimeError(f"Bulk remove failed for interfaces {self._pending_removes}: {e}") from e
-
-    def _remove_interfaces(self) -> ResponseType:
-        """
-        # Summary
-
-        Remove queued interfaces via `interfaceActions/remove`. Sends the explicit list of `{interfaceName, switchId}` pairs.
-
-        ## Raises
-
-        ### Exception
-
-        - If the remove API request fails (propagated to caller).
-        """
-        api_endpoint = EpManageInterfacesRemove()
-        api_endpoint.fabric_name = self.fabric_name
-        payload = {"interfaces": [{"interfaceName": name, "switchId": switch_id} for name, switch_id in self._pending_removes]}
-        return self.sender.request(path=api_endpoint.path, method=api_endpoint.verb, data=payload)
 
     def create(self, model_instance: LoopbackInterfaceModel, **kwargs) -> ResponseType:
         """
